+++ WBooks/Book_Details/BookDetailsViewController.swift
@@ -10,12 +10,6 @@ import UIKit
 
 class BookDetailsViewController: UIViewController {
     
-    // MARK: - Constants
-    
-    struct Constants {
-        static let headerImageHeight: CGFloat = 30
-    }
-    
     // MARK: - Properties
     
     private lazy var _view: BookDetailsView = BookDetailsView.loadFromNib()!
@@ -41,32 +35,12 @@ class BookDetailsViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        addHeaderImage()
+        addHeaderImage(to: self.view)
         
         load(childViewController: _bookInformationViewController,
              into: _view.informationContainerView,
-             with: UIEdgeInsets(top: 0,
-                                left: 0,
-                                bottom: 0,
-                                right: 0),
              respectSafeArea: true)
         
         _view.informationHeigthConstraint.constant = _bookInformationViewController.getViewHeigth()
     }
-    
-    // MARK: - Private Methods
-    
-    private func addHeaderImage() {
-        let headerImage = UIImageView(image: GeneralConstants.Design.navigationBarImage)
-        headerImage.contentMode = .bottom
-        headerImage.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(headerImage)
-        
-        NSLayoutConstraint.activate([headerImage.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-                                     headerImage.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-                                     headerImage.topAnchor.constraint(equalTo: view.topAnchor),
-                                     headerImage.heightAnchor.constraint(equalToConstant: Constants.headerImageHeight)
-            ])
-    }
-
 }
