+++ WBooks/Helpers/GeneralConstants.swift
@@ -16,5 +16,6 @@ struct GeneralConstants {
         static let navigationBarTitleFont = UIFont(name: "SFUIText-Semibold", size: 17)!
         static let navigationBarImage = UIImage(named: "bc_nav bar")!
         static let appDefaultImage = UIImage(named: "default_image")!
+        static let headerImageHeight: CGFloat = 30
     }
 }
