+++ WBooks/Models/AuthUser.swift
@@ -11,7 +11,7 @@ import Networking
 
 struct AuthUser: AuthenticableUser {
     // swiftlint:disable:next line_length
-    let sessionToken: String? = "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxNDksInZlcmlmaWNhdGlvbl9jb2RlIjoiZkNmQWJ1bmpfS0xmVkp6cjhIeWQ3dHc1VlJSa1ZNQi1iNFhZWnlHd0pYc3NMenpUVm1uY3RHLTVGcFBxRHhocyIsInJlbmV3X2lkIjoiQjlweG0xUE5qa1F5VGcxNDd6VmJ5dzZ6S2Rzc193X3JGeDN5dWFZXzdtQm9nNTdKTnVOdG9QemVvRVR0c2N4QiIsIm1heGltdW1fdXNlZnVsX2RhdGUiOjE1Mzg1ODI5NjcsImV4cGlyYXRpb25fZGF0ZSI6MTUzNjE2Mzc2Nywid2FybmluZ19leHBpcmF0aW9uX2RhdGUiOjE1MzYwMDg5Njd9.XZP9rZ1So6LAmViGoda6IIViWP8gpeoedHbHNsCSbFk"
+    let sessionToken: String? = "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxNDksInZlcmlmaWNhdGlvbl9jb2RlIjoiZkNmQWJ1bmpfS0xmVkp6cjhIeWQ3dHc1VlJSa1ZNQi1iNFhZWnlHd0pYc3NMenpUVm1uY3RHLTVGcFBxRHhocyIsInJlbmV3X2lkIjoieTJkMzRuUE1oVTZ6TVlxQVZVTGQ0dm1telFCZFNablF2M21LeGRiYjVISzZRdjY4UFlmLTdVNlBwakdyc1AzbiIsIm1heGltdW1fdXNlZnVsX2RhdGUiOjE1Mzg3NTc3ODEsImV4cGlyYXRpb25fZGF0ZSI6MTUzNjMzODU4MSwid2FybmluZ19leHBpcmF0aW9uX2RhdGUiOjE1MzYxODM3ODF9.6clJojop7TLcpfImhsbYgNj8L8zEbTlPL6RlVsCqoLo"
 }
 
 /*
