+++ WBooks/Repositories/BooksRepository.swift
@@ -34,13 +34,14 @@ class BooksRepository: AbstractRepository, BooksRepositoryType {
     func getBookStatus(id: Int) -> SignalProducer<BookStatus, RepositoryError> {
         let path = Constants.booksList + "/\(id)/rents"
         let result: SignalProducer<[Rent], RepositoryError> = performRequest(method: .get, path: path) { decode($0).toResult() }
+        
         let resultBool: SignalProducer<Bool, RepositoryError> = result.map({ rents in
             let rent = rents.filter({ (rent) in
                 rent.returnedDate != .none
             }).first
             return rent != nil
         })
-        
+    
         let resultStatus: SignalProducer<BookStatus, RepositoryError> = resultBool.map({ resBool in
             if resBool {
                 return BookStatus.avaliable
